Modificación a módulos, controlador y servicio de Categorías

+++ src/categories/categories.service.ts
@@ -1,20 +1,34 @@
-import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
+import { ConflictException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Category } from './entities/category.entity';
-import { Not, Repository } from 'typeorm';
+import { Movie } from 'src/movies/entities/movie.entity';
+import { Repository } from 'typeorm';
 
 @Injectable()
 export class CategoriesService {
   constructor(
     @InjectRepository(Category)
     private categoriesRepository: Repository<Category>,
+    @InjectRepository(Movie)
+    private moviesRepository: Repository<Movie>,
   ) {}
 
-  private async findOneOrFail(id: number): Promise<Category> {
+  private async findOneOrFail(id: number, relations = false): Promise<Category> {
     const category = await this.categoriesRepository.findOne({
-      where: { id }
+      where: { id: id },
+      relations: {
+        movies: (relations === true ? true : false)
+      },
+      select: {
+        movies: {
+          id: true,
+          title: true,
+          releaseDate: true,
+          posterUrl: true
+        }
+      }
     });
     if (!category) {
       throw new NotFoundException(`La categoría con el Id ${id} no existe`);
@@ -22,7 +36,8 @@ export class CategoriesService {
     return category;
   }
 
-  async create(createCategoryDto: CreateCategoryDto) {
+  async create(createCategoryDto: CreateCategoryDto): Promise<Category> {
+
     const existsCategory = await this.categoriesRepository.exists({
       where: { 
         title: createCategoryDto.title
@@ -34,28 +49,43 @@ export class CategoriesService {
     return this.categoriesRepository.save(createCategoryDto);
   }
 
-  findAll() {
-    return this.categoriesRepository.find();
+  async findAll(page = 1, limit = 10, relations = false): Promise<{ data: Category[]; total: number; page: number; limit: number }>  {
+    const [data, total] = await this.categoriesRepository.findAndCount({
+      skip: page > 0 ? (page - 1) * limit : 0,
+      take: limit,
+      select: {
+        id: true,
+        title: true,
+        description: true,
+        movies: {
+          id: true,
+          title: true,
+          releaseDate: true,
+          posterUrl: true
+        }
+      },
+      relations: {
+        movies: relations ? true : false
+      }
+    });
+  
+    return {
+      data,
+      total,
+      page,
+      limit,
+    };
   }
 
-  findOne(id: number) {
-    return this.findOneOrFail(id);
+  findOne(id: number, relations: boolean): Promise<Category> {
+    return this.findOneOrFail(id, relations);
   }
 
-  async update(id: number, updateCategoryDto: UpdateCategoryDto) {
+  async update(id: number, updateCategoryDto: UpdateCategoryDto): Promise<Category> {
+
     const category = await this.findOneOrFail(id);
 
     if (updateCategoryDto.title != null) {
-      const existsCategory = await this.categoriesRepository.exists({
-        where: { 
-          title: updateCategoryDto.title,
-          id: Not(id)
-        }
-      });
-      if (existsCategory) {
-        throw new ConflictException('El título ya está registrado');
-      }
-
       category.title = updateCategoryDto.title;
     }
     if (updateCategoryDto.description != null) {
@@ -68,9 +98,23 @@ export class CategoriesService {
     return this.categoriesRepository.save(category);
   }
 
-  async remove(id: number) {
+  async remove(id: number, cascade: boolean) {
     const category = await this.findOneOrFail(id);
+    if (cascade) {
+      await this.moviesRepository.delete({ category });
+      return this.categoriesRepository.delete(id);
+    } else {
+      const countMovies = await this.moviesRepository.countBy({ categoryId: id });
+      if (countMovies > 0) {
+        throw new ConflictException({ message: `La categoría no se puede eliminar porque tiene ${countMovies} películas asociadas` })
+      } else {
+        return this.categoriesRepository.delete(id);
+      }
+    }
+  }
 
-    return this.categoriesRepository.delete(id);
+  async findMovies(id: number): Promise<Movie[]> {
+    const category = await this.findOneOrFail(id, true);
+    return category.movies;
   }
 }
